% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_sheet_list}
\alias{as_sheet_list}
\title{Prepare a data frame or list of data frames for inclusion in a workbook}
\usage{
as_sheet_list(
  x,
  sheet_names = NULL,
  ...,
  n_sheets = NULL,
  .prep_fn = prep_wb_data,
  repair = "unique",
  default = "Sheet",
  call = caller_env()
)
}
\arguments{
\item{sheet_names}{One or more sheet names. Empty values ("" or " ") are
replaced using the \code{default} parameter value and position in \code{sheet_names}
vector.}

\item{...}{
  Arguments passed on to \code{\link{prep_wb_data}}
  \describe{
    \item{\code{x}}{Required. A data frame or an object coercible to a data frame with
\code{\link[base:as.data.frame]{base::as.data.frame()}}.}
    \item{\code{list_columns}}{String, one of "collapse" (default), "drop", or "asis"}
    \item{\code{sep}}{String to use in collapsing list columns. Ignored unless
\code{list_columns = "collapse"}. Defaults to \code{"; "}.}
    \item{\code{geometry}}{String, one of "drop" (default), "coords", or "wkt". "coords"
uses \code{\link[sf:st_centroid]{sf::st_centroid()}} to convert input to POINT geometry, transforms
geometry to EPSG:4326, converts geometry to coordinates, and adds new
columns with names matching \code{coords}. "wkt" converts geometry to a Well
Known Text (WKT) character vector using \code{\link[sf:st_as_text]{sf::st_as_text()}} and replaces the
existing geometry column (keeping the existing sf column name).}
    \item{\code{coords}}{Length 2 character vector with column names to add if \code{geometry = "coords"}. Must be length 2 in longitude, latitude order.}
    \item{\code{call}}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
  }}

\item{n_sheets}{Number of sheet names to return.}

\item{.prep_fn}{Function to use in preparing list of input objects. Defaults
to \code{\link[=prep_wb_data]{prep_wb_data()}}.}

\item{default}{Default prefix to use for numbered sheets. Default values are
used if \code{sheet_names = NULL} or if \code{n_sheets} is greater than the length of
\code{sheet_names}. Defaults to "Sheet".}
}
\value{
A named list of data frames.
}
\description{
\code{\link[=as_sheet_list]{as_sheet_list()}} prepares a list of data frames for inclusion in a workbook.
}
\examples{

x <- head(mtcars, 1)
y <- tail(mtcars, 1)

as_sheet_list(x)

as_sheet_list(list(x, y))

as_sheet_list(list(head = x, tail = y))

as_sheet_list(list(head = x, y))

}
\keyword{internal}
\keyword{utils}
