% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{common_dxfs_styles}
\alias{common_dxfs_styles}
\title{Common dxfs styles}
\source{
https://stackoverflow.com/questions/27611260/what-are-the-rgb-codes-for-the-conditional-formatting-styles-in-excel#comment78058968_27611522
}
\usage{
common_dxfs_styles
}
\description{
\code{common_dxfs_styles} is a named list of dxfs styles created with
\code{\link[openxlsx2:create_dxfs_style]{openxlsx2::create_dxfs_style()}} with bad, good, and neutral values.
}
\seealso{
\code{\link[=wb_add_styles]{wb_add_styles()}}
}
