% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marquee.R
\name{fmt_marquee_txt}
\alias{fmt_marquee_txt}
\title{Use \code{marquee::marquee_parse()} to format Markdown text}
\usage{
fmt_marquee_txt(
  text,
  ...,
  style = marquee::classic_style(),
  ignore_html = TRUE
)
}
\arguments{
\item{text}{A character string. The core quality of markdown is that any text
is valid markdown so there is no restrictions on the content}

\item{...}{Ignored at this time. For future use.}

\item{style}{A style set such as \code{\link[marquee:classic_style]{classic_style()}} that defines how the text
should be rendered}

\item{ignore_html}{Should HTML code be removed from the output}
}
\value{
A list of \code{fmt_txt} objects, one per parsed line of \code{text}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{\link[=fmt_marquee_txt]{fmt_marquee_txt()}} uses \code{\link[marquee:marquee_parse]{marquee::marquee_parse()}} and
\code{\link[openxlsx2:fmt_txt]{openxlsx2::fmt_txt()}} to format Markdown text.
}
\examples{

fmt_marquee_txt(
  "# ABC

  abc"
)

}
