% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magic_conversion_functions.R
\name{magic_conversion_functions}
\alias{magic_conversion_functions}
\alias{csv_to_wb}
\alias{csv_to_xlsx}
\alias{xlsx_to_csv}
\title{Magic Conversion Functions}
\usage{
csv_to_wb(csv, fun = utils::read.csv, ...)

csv_to_xlsx(csv, xlsx, fun = utils::read.csv, ...)

xlsx_to_csv(xlsx, csv, sheet = 1, fun = utils::write.csv, ...)
}
\arguments{
\item{csv, xlsx}{name of input and output files. It is possible to pass
multiple CSV files when creating a workbook or XLSX file. If these inputs
are named, the names will be used as worksheet names.}

\item{fun}{a function used to read or write the csv file. Defaults to
\code{\link[utils:read.table]{utils::read.csv()}} and \code{\link[utils:write.table]{utils::write.csv()}}, but other functions from other
packages can be used too}

\item{...}{additional arguments passed to \code{fun}}

\item{sheet}{a sheet in the workbook (either an index or a sheet name)}
}
\description{
A set of functions to convert between CSV and XLSX formats using
flexible input and output options.
}
\details{
These functions allow seamless conversion between CSV and XLSX formats.
\itemize{
\item \code{csv_to_wb}: Reads one or more CSV files and writes them to a workbook
object.
\item \code{csv_to_xlsx}: Converts CSV files to an XLSX file.
\item \code{xlsx_to_csv}: Converts an XLSX file to a CSV file.
}
}
\examples{
# Create example CSV file
csv <- tempfile(fileext = ".csv")
utils::write.csv(x = mtcars, file = csv)

# Convert CSV to Workbook
wb <- csv_to_wb(csv = csv)

# Convert CSV to XLSX
xlsx <- openxlsx2::temp_xlsx()
csv_to_xlsx(csv = csv, xlsx = xlsx)

# Convert XLSX back to CSV
xlsx_to_csv(xlsx = xlsx, csv = csv)
}
