% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_wb.R
\name{map_wb}
\alias{map_wb}
\title{Convert an object to a list of workbooks}
\usage{
map_wb(x, ..., properties = NULL, .progress = FALSE)
}
\arguments{
\item{x}{A data frame, a list of data frames, a file path for an Excel file,
or a \code{wbWorkbook} object. A \code{wbWorkbook} is returned "as is" ignoring all
other parameters. A file path is loaded to a data frame using
\code{\link[openxlsx2:wb_to_df]{openxlsx2::wb_to_df()}}.}

\item{...}{
  Arguments passed on to \code{\link{as_wb}}
  \describe{
    \item{\code{sheet_names}}{Optional character vector of worksheet names.}
    \item{\code{title,subject,category,keywords}}{Additional arguments passed to
\code{\link[openxlsx2:wb_workbook]{openxlsx2::wb_workbook()}}.}
    \item{\code{creator}}{Creator of the workbook (your name). Defaults to login username or \code{options("openxlsx2.creator")} if set.}
    \item{\code{datetime_created}}{The time of the workbook is created}
    \item{\code{theme}}{Optional theme identified by string or number.
See \strong{Details} for options.}
    \item{\code{call}}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
  }}

\item{properties}{A named list (typically from
\code{\link[openxlsx2:wb_get_properties]{openxlsx2::wb_get_properties()}}) used to set new workbook properties for
any values set to \code{NULL}. \code{datetime_created} defaults to \code{\link[=Sys.time]{Sys.time()}} so
must be set to \code{NULL} to inherit value from \code{properties}.}

\item{.progress}{Whether to show a progress bar. Use \code{TRUE} to turn on
a basic progress bar, use a string to give it a name, or see
\link[purrr]{progress_bars} for more details.}
}
\value{
A list of wbWorkbook objects.
}
\description{
\code{\link[=map_wb]{map_wb()}} takes a list and returns a list of \code{wbWorkbook} objects.
\code{properties} is recycled to match the length of the input x.
}
\examples{
map_wb(list(mtcars[1:3, ], mtcars[4:6, ]))

}
\keyword{internal}
