% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt.R
\name{set_excel_fmt_class}
\alias{set_excel_fmt_class}
\title{Prepare a data frame with Excel style class values for formatting by
openxlsx2}
\usage{
set_excel_fmt_class(
  data,
  cols,
  fmt_class = "currency",
  multiple = TRUE,
  strict = FALSE
)
}
\arguments{
\item{data}{A data frame with columns to format.}

\item{cols}{Column names or numbers to modify.}

\item{fmt_class}{Excel style class, one of: c("currency", "accounting",
"hyperlink", "percentage", "scientific", "formula"). Length is recycled to
match length of cols using \code{\link[vctrs:vec_recycle]{vctrs::vec_recycle()}}.}

\item{multiple}{Whether \code{arg} may contain zero or several values.}

\item{strict}{If \code{TRUE}, error if any character values in cols are not
included in the names of \code{data}. Default \code{FALSE}.}
}
\value{
A data frame matching \code{data} with \code{fmt_class} added to the class
of the specified \code{cols}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
\code{\link[=set_excel_fmt_class]{set_excel_fmt_class()}} applies a style to each specified column. See the
openxlsx2 documentation for more information on this feature:
\url{https://janmarvin.github.io/openxlsx2/articles/openxlsx2_style_manual.html#numfmts2}
}
