% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_lgl_cols.R
\name{vec_fmt_lgl}
\alias{vec_fmt_lgl}
\title{Format a logical vector to specified values}
\usage{
vec_fmt_lgl(x, values = c("Y", "N"))
}
\arguments{
\item{x}{A logical vector.}

\item{values}{Default to \code{c("Y", "N")} Length 2 vector where first element
replaces \code{TRUE} and the second element replaces \code{FALSE}.}
}
\value{
A character vector the same length as \code{x}.
}
\description{
Format a logical vector to specified values
}
\examples{
vec_fmt_lgl(c(TRUE, FALSE))

vec_fmt_lgl(c(TRUE, FALSE), c("Yes", "No"))
}
