% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{wb_add_styles}
\alias{wb_add_styles}
\title{Add a named list of styles to a workbook}
\usage{
wb_add_styles(wb, styles, style_names = NULL)
}
\arguments{
\item{wb}{A \link[openxlsx2]{wbWorkbook} object.}

\item{styles}{Required. List or style xml character, created by a
⁠create_*()⁠ function. Passed to \code{style} argument of
\code{\link[openxlsx2:wb_add_style]{openxlsx2::wb_add_style()}}.}

\item{style_names}{Optional if styles is a named list.}
}
\description{
\code{\link[=wb_add_styles]{wb_add_styles()}} extends \code{\link[openxlsx2:wb_add_style]{openxlsx2::wb_add_style()}} by allowing the
addition of multiple styles as a named list.
}
