% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dims.R
\name{wb_dims_ext}
\alias{wb_dims_ext}
\title{Extended helper to specify the dims or cols arguments}
\usage{
wb_dims_ext(
  wb = NULL,
  sheet = openxlsx2::current_sheet(),
  cols,
  x = NULL,
  select = NULL,
  start_row = NULL,
  start_col = NULL,
  error_call = rlang::caller_env()
)
}
\arguments{
\item{select}{A string, one of the followings.
it improves the selection of various parts of \code{x}
One of "x", "data", "col_names", or "row_names".
\code{"data"} will only select the data part, excluding row names and column names  (default if \code{cols} or \code{rows} are specified)
\code{"x"} Includes the data, column and row names if they are present. (default if none of \code{rows} and \code{cols} are provided)
\code{"col_names"} will only return column names
\code{"row_names"} Will only return row names.}
}
\value{
A dims character string, or (if \code{select = "cols"}) an integer
vector of the selected column positions.
}
\description{
\code{\link[=wb_dims_ext]{wb_dims_ext()}} extends \code{\link[openxlsx2:wb_dims]{openxlsx2::wb_dims()}} by allowing a workbook input
used to set \code{x} and supporting tidy selection for the \code{cols} argument.
}
\examples{
wb <- as_wb(mtcars[1:3, ])

wb_dims_ext(wb, sheet = "Sheet 1", cols = c(mpg, hp))

}
\keyword{internal}
