% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_workbook.R
\name{wb_new_workbook}
\alias{wb_new_workbook}
\title{Create a new workbook and add named work sheets}
\usage{
wb_new_workbook(
  sheet_names = NULL,
  ...,
  default = "Sheet",
  creator = NULL,
  title = NULL,
  subject = NULL,
  category = NULL,
  datetime_created = Sys.time(),
  theme = NULL,
  keywords = NULL,
  properties = NULL,
  call = caller_env()
)
}
\arguments{
\item{sheet_names}{Optional character vector of worksheet names.}

\item{...}{
  Arguments passed on to \code{\link[openxlsx2:wb_add_worksheet]{openxlsx2::wb_add_worksheet}}
  \describe{
    \item{\code{grid_lines}}{Logical; if \code{FALSE}, the worksheet grid lines are hidden.}
    \item{\code{row_col_headers}}{Logical; if \code{FALSE}, row numbers and column letters are hidden.}
    \item{\code{tab_color}}{The color of the worksheet tab. Accepts a \code{\link[openxlsx2:wb_color]{wb_color()}} object,
a standard R color name, or a hex color code (e.g., "#4F81BD").}
    \item{\code{zoom}}{The sheet zoom level as a percentage; a numeric value between
10 and 400. Values below 10 default to 10.}
    \item{\code{header,footer}}{Default character vectors of length three for the left,
center, and right sections of the header or footer.}
    \item{\code{odd_header,odd_footer}}{Specific definitions for odd-numbered pages.
Defaults to the values provided in \code{header} and \code{footer}.}
    \item{\code{even_header,even_footer}}{Specific definitions for even-numbered pages.
Defaults to the values provided in \code{header} and \code{footer}.}
    \item{\code{first_header,first_footer}}{Specific definitions for the first page
of the worksheet. Defaults to the values provided in \code{header} and \code{footer}.}
    \item{\code{visible}}{The visibility state of the sheet. One of "visible", "hidden",
or "veryHidden".}
    \item{\code{has_drawing}}{\emph{defunct}}
    \item{\code{paper_size}}{An integer code representing a standard paper size.
Refer to \code{\link[openxlsx2:wb_page_setup]{wb_page_setup()}} for a complete list of codes.}
    \item{\code{orientation}}{The page orientation, either "portrait" or "landscape".}
    \item{\code{hdpi,vdpi}}{The horizontal and vertical DPI (dots per inch) for
printing and rendering. Can be set globally via \code{options("openxlsx2.hdpi")}.}
  }}

\item{default}{Default prefix to use for numbered sheets. Default values are
used if \code{sheet_names = NULL} or if \code{n_sheets} is greater than the length of
\code{sheet_names}. Defaults to "Sheet".}

\item{creator}{Creator of the workbook (your name). Defaults to login username or \code{options("openxlsx2.creator")} if set.}

\item{title, subject, category, keywords}{Additional arguments passed to
\code{\link[openxlsx2:wb_workbook]{openxlsx2::wb_workbook()}}.}

\item{datetime_created}{The time of the workbook is created}

\item{theme}{Optional theme identified by string or number.
See \strong{Details} for options.}

\item{properties}{A named list (typically from
\code{\link[openxlsx2:wb_get_properties]{openxlsx2::wb_get_properties()}}) used to set new workbook properties for
any values set to \code{NULL}. \code{datetime_created} defaults to \code{\link[=Sys.time]{Sys.time()}} so
must be set to \code{NULL} to inherit value from \code{properties}.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
A \code{wbWorkbook} object.
}
\description{
\code{\link[=wb_new_workbook]{wb_new_workbook()}} is a convenience function wrapping
\code{\link[openxlsx2:wb_workbook]{openxlsx2::wb_workbook()}} and \code{\link[openxlsx2:wb_add_worksheet]{openxlsx2::wb_add_worksheet()}} to offer quick
and easy setup for new workbooks.
}
\examples{
wb_new_workbook()

wb_new_workbook("Sheet 1")

wb_new_workbook(c("Data", "Analysis"))

}
\seealso{
\code{\link[=as_wb]{as_wb()}}
}
