% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_rename.R
\name{wb_rename_sheets}
\alias{wb_rename_sheets}
\alias{wb_rename_sheets_with}
\title{Rename workbook sheets using a tidyselect syntax}
\usage{
wb_rename_sheets(wb, ...)

wb_rename_sheets_with(wb, .fn, .sheets = tidyselect::everything())
}
\arguments{
\item{wb}{A \code{wbWorkbook} object.}

\item{...}{For \code{wb_rename_sheets()}, \verb{<new_name> = <old_name>} pairs passed
to \code{\link[tidyselect:eval_rename]{tidyselect::eval_rename()}} specifying how to rename sheets.}

\item{.fn}{Function used to transform the selected sheet names.}

\item{.sheets}{Tidyselect expression selecting which sheets to rename.
Defaults to all sheets.}
}
\value{
A \code{wbWorkbook} object.

A \code{wbWorkbook} object.
}
\description{
\code{wb_rename_sheets()} and \code{wb_rename_sheets_with()} use the tidyselect package
to rename sheets.
}
\examples{
wb <- as_wb(list(a = mtcars[1:2, ], b = mtcars[3:4, ]))
wb_rename_sheets(wb, sales = "a", inventory = "b")

wb2 <- as_wb(list(a = mtcars[1:2, ], b = mtcars[3:4, ]))
wb_rename_sheets_with(wb2, \(x) paste0("sheet_", x))

}
