% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_set_col_widths_ext.R
\name{wb_set_col_widths_ext}
\alias{wb_set_col_widths_ext}
\title{Set workbook column widths}
\usage{
wb_set_col_widths_ext(
  wb,
  cols = NULL,
  widths = "auto",
  sheet = current_sheet()
)
}
\arguments{
\item{wb}{A \code{wbWorkbook} object.}

\item{cols}{Indices of cols to set/remove column widths.}

\item{widths}{Width to set \code{cols} to specified column width or \code{"auto"} for
automatic sizing. \code{widths} is recycled to the length of \code{cols}. openxlsx2
sets the default width is 8.43, as this is the standard in some spreadsheet
software. See \strong{Details} for general information on column widths.}

\item{sheet}{A name or index of a worksheet, a vector in the case of \code{remove_}}
}
\value{
A \code{wbWorkbook} object.
}
\description{
\code{\link[=wb_set_col_widths_ext]{wb_set_col_widths_ext()}} extends \code{\link[openxlsx2:wb_set_col_widths]{openxlsx2::wb_set_col_widths()}} by making
\code{cols} optional (defaults to all columns) and allows setting columns based on
column names (rather than index position alone).
}
\examples{
wb <- as_wb(mtcars)

wb_set_col_widths_ext(wb, cols = "drat")

# `cols` is optional
wb_set_col_widths_ext(wb)

}
\seealso{
\code{\link[openxlsx2:wb_data]{openxlsx2::wb_data()}}, \code{\link[openxlsx2:col_widths-wb]{openxlsx2::col_widths-wb()}}
}
\keyword{internal}
