% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{as_data_validation_value}
\alias{as_data_validation_value}
\title{Prepare a character vector as a data validation string}
\usage{
as_data_validation_value(x, allow_blank = TRUE)
}
\arguments{
\item{x}{A character vector of options to allow in cell validation.}

\item{allow_blank}{If \code{TRUE}, add a blank space \code{" "} to the returned values.}
}
\description{
\code{\link[=as_data_validation_value]{as_data_validation_value()}} collapses a character vector as a single string
for use as a validation \code{value} argument by
\code{\link[openxlsx2:wb_add_data_validation]{openxlsx2::wb_add_data_validation()}} when \code{type = "list"}.
}
