% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_wb.R
\name{as_wb}
\alias{as_wb}
\title{Coerce a data frame or list of data frames to a workbook}
\usage{
as_wb(
  x,
  ...,
  sheet_names = NULL,
  creator = NULL,
  title = NULL,
  subject = NULL,
  category = NULL,
  datetime_created = Sys.time(),
  theme = NULL,
  keywords = NULL,
  type = c("df-list", "df", "any"),
  call = caller_env()
)
}
\arguments{
\item{x}{A data frame or list of data frames. Objects that can be coerced to
a data frame are not supported.}

\item{...}{
  Arguments passed on to \code{\link[=wb_add_data_ext]{wb_add_data_ext}}
  \describe{
    \item{\code{as_table}}{Default \code{FALSE}. If \code{TRUE}, use
\code{\link[openxlsx2:wb_add_data_table]{openxlsx2::wb_add_data_table()}} to add data to workbook. If \code{FALSE}, use
\code{\link[openxlsx2:wb_add_data]{openxlsx2::wb_add_data()}}. Additional parameters in \code{...} are passed to
one function or the other depending on this value.}
    \item{\code{labels}}{Method for handling column labels. "drop" (default) or
"row_before". If "row_before", insert column labels in the row before the
column names.}
    \item{\code{list_columns}}{String, one of "collapse" (default), "drop", or "asis"}
    \item{\code{sep}}{String to use in collapsing list columns. Ignored unless
\code{list_columns = "collapse"}. Defaults to \code{"; "}.}
    \item{\code{geometry}}{String, one of "drop" (default), "coords", or "wkt". "coords"
uses \code{\link[sf:geos_unary]{sf::st_centroid()}} to convert input to POINT geometry, transforms
geometry to EPSG:4326, converts geometry to coordinates, and adds new
columns with names matching \code{coords}. "wkt" converts geometry to a Well
Known Text (WKT) character vector using \code{\link[sf:st_as_text]{sf::st_as_text()}} and replaces the
existing geometry column (keeping the existing sf column name).}
    \item{\code{coords}}{Length 2 character vector with column names to add if \code{geometry = "coords"}. Must be length 2 in longitude, latitude order.}
    \item{\code{wb}}{A Workbook object containing a worksheet.}
    \item{\code{sheet}}{The worksheet to write to. Can be the worksheet index or name.}
    \item{\code{start_row}}{A vector specifying the starting row to write \code{x} to.}
  }}

\item{sheet_names}{Optional character vector of worksheet names.}

\item{creator}{Creator of the workbook (your name). Defaults to login username or \code{options("openxlsx2.creator")} if set.}

\item{title, subject, category, keywords}{Additional arguments passed to
\code{\link[openxlsx2:wb_workbook]{openxlsx2::wb_workbook()}}.}

\item{datetime_created}{The time of the workbook is created}

\item{theme}{Optional theme identified by string or number.
See \strong{Details} for options.}

\item{type}{Type of objects to allow. "df-list" allows data frames and lists
of data frames. "df" allows data frames only. "any" allows any input
(allowing the option for \code{\link[=wb_add_data_ext]{wb_add_data_ext()}} to coerce objects to data
frames).}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
\code{\link[=df_to_wb]{df_to_wb()}} converts a data frame or list of data frames to a wbWorkbook
object.
}
\examples{
df_to_wb(mtcars)

df_to_wb(list(mtcars, mtcars))

}
\seealso{
\code{\link[=write_excel_ext]{write_excel_ext()}}
}
