% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_lgl_cols.R
\name{fmt_lgl_cols}
\alias{fmt_lgl_cols}
\title{Format logical vector columns to use replacement values}
\usage{
fmt_lgl_cols(
  .data,
  .cols = tidyselect::where(is.logical),
  values = c("Y", "N")
)
}
\arguments{
\item{.cols}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Columns to transform.
You can't select grouping columns because they are already automatically
handled by the verb (i.e. \code{\link[dplyr:summarise]{summarise()}} or \code{\link[dplyr:mutate]{mutate()}}).}

\item{values}{Default to \code{c("Y", "N")} Length 2 vector where first element
replaces \code{TRUE} and the second element replaces \code{FALSE}.}
}
\description{
\code{\link[=fmt_lgl_cols]{fmt_lgl_cols()}} uses \code{\link[=vec_fmt_lgl]{vec_fmt_lgl()}} to format all (or specified) logical
vector columns to use replacement values.
}
\examples{
fmt_lgl_cols(data.frame(x = c(TRUE, FALSE, TRUE)))
}
