% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openxlsx2Extras-package.R
\docType{package}
\name{openxlsx2Extras-package}
\alias{openxlsx2Extras}
\alias{openxlsx2Extras-package}
\title{openxlsx2Extras: Extra Functions for the openxlsx2 package}
\description{
Extends the openxlsx2 package with wrapper and helper functions designed to add new features and options when working with Excel workbooks.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/elipousson/openxlsx2Extras}
  \item \url{https://elipousson.github.io/openxlsx2Extras/}
  \item Report bugs at \url{https://github.com/elipousson/openxlsx2Extras/issues}
}

}
\author{
\strong{Maintainer}: Eli Pousson \email{eli.pousson@gmail.com} (\href{https://orcid.org/0000-0001-8280-1706}{ORCID}) [copyright holder]

}
\keyword{internal}
