% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_xlsx_ext.R
\name{read_xlsx_ext}
\alias{read_xlsx_ext}
\title{Create a data frame from a Workbook (with extra features)}
\usage{
read_xlsx_ext(
  file,
  sheet = 1,
  ...,
  names_from = "basename",
  names_to = rlang::zap(),
  combine = TRUE,
  repair = "unique_quiet"
)
}
\arguments{
\item{file}{A workbook file path, a \link[openxlsx2]{wbWorkbook} object, or a valid URL.}

\item{sheet}{Defaults to 1.}

\item{...}{
  Arguments passed on to \code{\link[openxlsx2:read_xlsx]{openxlsx2::read_xlsx}}
  \describe{
    \item{\code{start_row,start_col}}{Optional numeric values specifying the first row or column
to begin data discovery.}
    \item{\code{row_names}}{Logical; if TRUE, uses the first column of the selection as row names.}
    \item{\code{col_names}}{Logical; if TRUE, uses the first row of the selection as column headers.}
    \item{\code{skip_empty_rows,skip_empty_cols}}{Logical; if TRUE, filters out rows or
columns containing only missing values.}
    \item{\code{skip_hidden_rows,skip_hidden_cols}}{Logical; if TRUE, excludes rows or
columns marked as hidden in the worksheet metadata.}
    \item{\code{rows,cols}}{Optional numeric vectors specifying the exact indices to read.}
    \item{\code{detect_dates}}{Logical; if TRUE, identifies date and datetime styles for conversion.}
    \item{\code{na}}{A character vector or a named list (e.g., \code{list(strings = "", numbers = -99)})
defining values to treat as \code{NA}.}
    \item{\code{fill_merged_cells}}{Logical; if TRUE, propagates the top-left value of a
merged range to all cells in that range.}
    \item{\code{dims}}{A character string defining the range. Supports wildcards
(e.g., "A1:++" or "A-:+5").}
    \item{\code{show_formula}}{Logical; if TRUE, returns the formula strings instead of
calculated values.}
    \item{\code{convert}}{Logical; if TRUE, attempts to coerce columns to appropriate R classes.}
    \item{\code{types}}{A named vector (numeric or character) to explicitly define column types.}
    \item{\code{named_region}}{A character string referring to a defined name or spreadsheet Table.}
    \item{\code{keep_attributes}}{Logical; if TRUE, attaches metadata such as the internal
type table (tt) and types as attributes to the output.}
    \item{\code{check_names}}{Logical; if TRUE, ensures column names are syntactically
valid R names via \code{\link[=make.names]{make.names()}}.}
    \item{\code{show_hyperlinks}}{Logical; if TRUE, replaces cell values with their
underlying hyperlink targets.}
    \item{\code{apply_numfmts}}{Logical; if TRUE, applies spreadsheet number formatting
and returns strings.}
  }}

\item{names_to}{By default, \code{names(x)} are lost. To keep them, supply a
string to \code{names_to} and the names will be saved into a column with that
name. If \code{names_to} is supplied and \code{x} is not named, the position of
the elements will be used instead of the names.}

\item{combine}{If \code{TRUE}, always return a data frames. If \code{FALSE}, return a
list of data frames.}

\item{repair}{Either a string or a function. If a string, it must be one of
\code{"check_unique"}, \code{"minimal"}, \code{"unique"}, \code{"universal"}, \code{"unique_quiet"},
or \code{"universal_quiet"}. If a function, it is invoked with a vector of
minimal names and must return minimal names, otherwise an error is thrown.
\itemize{
\item Minimal names are never \code{NULL} or \code{NA}. When an element doesn't
have a name, its minimal name is an empty string.
\item Unique names are unique. A suffix is appended to duplicate
names to make them unique.
\item Universal names are unique and syntactic, meaning that you can
safely use the names as variables without causing a syntax
error.
}

The \code{"check_unique"} option doesn't perform any name repair.
Instead, an error is raised if the names don't suit the
\code{"unique"} criteria.

The options \code{"unique_quiet"} and \code{"universal_quiet"} are here to help the
user who calls this function indirectly, via another function which exposes
\code{repair} but not \code{quiet}. Specifying \code{repair = "unique_quiet"} is like
specifying \verb{repair = "unique", quiet = TRUE}. When the \code{"*_quiet"} options
are used, any setting of \code{quiet} is silently overridden.}
}
\value{
A data frame, or (if \code{combine = FALSE}) a list of data frames.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{\link[=read_xlsx_ext]{read_xlsx_ext()}} uses \code{\link[openxlsx2:read_xlsx]{openxlsx2::read_xlsx()}} with a few added features:
\itemize{
\item allows use of a name repair argument (\code{"unique_quite"} by default) to avoid
blank \code{""} or \code{NA} values for column names.
\item allows vector inputs for the file or sheet argument. These parameters are
recycled to a common length and result in the return of a data frame list
unless the \code{combine = TRUE} is set. If \code{combine = TRUE}, set \code{names_to}
(passed to \code{purrr::list_rbind}) to combined the file basename values
(default) or full path values as a column (depending on the \code{names_from}
argument). \code{names_from} can also be a length > 1 character vector that can be
recycled to match the length of file.
}
}
