% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_sheet_list_names}
\alias{set_sheet_list_names}
\alias{as_sheet_names}
\title{Set sheet names for input data}
\source{
\url{https://github.com/JanMarvin/openxlsx2/blob/main/R/write_xlsx.R#L281-L301}
}
\usage{
set_sheet_list_names(
  x,
  sheet_names = NULL,
  ...,
  n_sheets = NULL,
  .prep_fn = prep_wb_data,
  repair = "unique",
  default = "Sheet",
  call = caller_env()
)

as_sheet_names(
  sheet_names = NULL,
  n_sheets = 1,
  default = "Sheet",
  max_length = 31,
  excess_length = c("truncate", "error"),
  repair = "unique",
  quiet = FALSE,
  arg = caller_arg(sheet_names),
  call = caller_env()
)
}
\arguments{
\item{sheet_names}{One or more sheet names. Empty values ("" or " ") are
replaced using the \code{default} parameter value and position in \code{sheet_names}
vector.}

\item{n_sheets}{Number of sheet names to return.}

\item{default}{Default prefix to use for numbered sheets. Default values are
used if \code{sheet_names = NULL} or if \code{n_sheets} is greater than the length of
\code{sheet_names}. Defaults to "Sheet".}

\item{max_length}{Maximum allowed length in characters for sheet names. Must
be 31 or less.}

\item{excess_length}{Handling for sheet names that exceed \code{max_length} in
characters. "truncate" trims to length of \code{max_length} and "error" errors.}
}
\description{
\code{\link[=set_sheet_list_names]{set_sheet_list_names()}} and \code{\link[=as_sheet_names]{as_sheet_names()}} are adapted from the existing
functionality for handling sheet names in \code{\link[openxlsx2:write_xlsx]{openxlsx2::write_xlsx()}}.
}
\examples{

as_sheet_names()

as_sheet_names("Workbook sheet")

as_sheet_names(c("Workbook sheet", "Workbook sheet"), quiet = TRUE)

as_sheet_names("Workbook sheet", n_sheets = 2)

as_sheet_names("Sheet names longer than 31 characters are truncated")

}
\keyword{internal}
\keyword{utils}
