% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_add.R
\name{wb_add_data_ext}
\alias{wb_add_data_ext}
\title{Add data using \code{openxlsx2::wb_add_data()} with extra features}
\usage{
wb_add_data_ext(
  wb,
  x,
  ...,
  start_row = 1,
  geometry = c("drop", "coords", "wkt"),
  labels = c("drop", "row_before"),
  call = caller_env()
)
}
\arguments{
\item{wb}{A Workbook object containing a worksheet.}

\item{x}{Object to be written. For classes supported look at the examples.}

\item{...}{
  Arguments passed on to \code{\link[openxlsx2:wb_add_data]{openxlsx2::wb_add_data}}
  \describe{
    \item{\code{sheet}}{The worksheet to write to. Can be the worksheet index or name.}
    \item{\code{dims}}{Spreadsheet cell range that will determine \code{start_col} and \code{start_row}: "A1", "A1:B2", "A:B"}
    \item{\code{start_col}}{A vector specifying the starting column to write \code{x} to.}
    \item{\code{array}}{A bool if the function written is of type array}
    \item{\code{col_names}}{If \code{TRUE}, column names of \code{x} are written.}
    \item{\code{row_names}}{If \code{TRUE}, the row names of \code{x} are written.}
    \item{\code{with_filter}}{If \code{TRUE}, add filters to the column name row.
NOTE: can only have one filter per worksheet.}
    \item{\code{name}}{The name of a named region if specified.}
    \item{\code{sep}}{Only applies to list columns. The separator used to collapse list
columns to a character vector e.g. \code{sapply(x$list_column, paste, collapse = sep)}.}
    \item{\code{apply_cell_style}}{Should we write cell styles to the workbook}
    \item{\code{remove_cell_style}}{keep the cell style?}
    \item{\code{na.strings}}{Value used for replacing \code{NA} values from \code{x}. Default
looks if \code{options(openxlsx2.na.strings)} is set. Otherwise \code{\link[openxlsx2:na_strings]{na_strings()}}
uses the special \verb{#N/A} value within the workbook.}
    \item{\code{inline_strings}}{write characters as inline strings}
    \item{\code{enforce}}{enforce that selected dims is filled. For this to work, \code{dims} must match \code{x}}
  }}

\item{start_row}{A vector specifying the starting row to write \code{x} to.}

\item{geometry}{String, one of "drop" (default), "coords", or "wkt". Both
"coords" and  "wkt" are not yet supported.}

\item{labels}{Method for handling column labels. "drop" (default) or
"row_before". If "row_before", insert column labels in the row before the
column names.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{\link[=wb_add_data_ext]{wb_add_data_ext()}} extends \code{\link[openxlsx2:wb_add_data]{openxlsx2::wb_add_data()}} to add data to a
workbook with special handling for sf input objects and labelled data.
}
