% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marquee.R
\name{wb_add_marquee_text}
\alias{wb_add_marquee_text}
\title{Add Markdown formatted text to a Workbook}
\usage{
wb_add_marquee_text(
  wb = NULL,
  text,
  sheet = current_sheet(),
  dims = NULL,
  ...,
  cols = 1,
  widths = "auto"
)
}
\arguments{
\item{wb}{A Workbook object containing a worksheet.}

\item{text}{A character string. The core quality of markdown is that any text
is valid markdown so there is no restrictions on the content}

\item{sheet}{The worksheet to write to. Can be the worksheet index or name.}

\item{dims}{Spreadsheet cell range that will determine \code{start_col} and \code{start_row}: "A1", "A1:B2", "A:B"}

\item{...}{
  Arguments passed on to \code{\link{fmt_marquee_txt}}
  \describe{
    \item{\code{style}}{A style set such as \code{\link[marquee:classic_style]{classic_style()}} that defines how the text
should be rendered}
    \item{\code{ignore_html}}{Should HTML code be removed from the output}
  }}

\item{cols}{Indices of cols to set/remove column widths.}

\item{widths}{Width to set \code{cols} to specified column width or \code{"auto"} for
automatic sizing. \code{widths} is recycled to the length of \code{cols}. openxlsx2
sets the default width is 8.43, as this is the standard in some spreadsheet
software. See \strong{Details} for general information on column widths.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{\link[=wb_add_marquee_text]{wb_add_marquee_text()}} uses \code{\link[=fmt_marquee_txt]{fmt_marquee_txt()}} to add Markdown formatted
text to a Workbook. \code{\link[openxlsx2:wb_set_col_widths]{openxlsx2::wb_set_col_widths()}} is applied to the
workbook to adjust widths for the specified columns. Set \code{widths = NULL} to
disable this functionality.
}
\examples{

library(openxlsx2)

wb <- wb_workbook()
wb <- wb_add_worksheet(wb)

wb <- wb_add_marquee_text(
  wb,
  text = "
# Heading 1

Example text.

~~Strikethrough text~~

## Heading 2

- Bulleted list item 1
  - Nested bullet
- Bulleted list item 2"
)

}
