% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{wb_add_styles}
\alias{wb_add_styles}
\title{Add a named list of styles to a workbook}
\usage{
wb_add_styles(wb, styles, style_names = NULL)
}
\arguments{
\item{wb}{A \link[openxlsx2]{wbWorkbook} object.}

\item{styles}{Required. List or style xml character, created by a
⁠create_*()⁠ function. Passed to \code{style} argument of
\code{\link[openxlsx2:wb_add_style]{openxlsx2::wb_add_style()}}.}

\item{style_names}{Optional if styles is a named list.}
}
\value{
A \code{wbWorkbook} object.
}
\description{
\code{\link[=wb_add_styles]{wb_add_styles()}} extends \code{\link[openxlsx2:wb_add_style]{openxlsx2::wb_add_style()}} by allowing the
addition of multiple styles as a named list.
}
\examples{
wb <- openxlsx2::wb_workbook()
wb <- openxlsx2::wb_add_worksheet(wb)

style <- openxlsx2::create_dxfs_style(font_color = openxlsx2::wb_color(hex = "FF0000"))
wb_add_styles(wb, styles = style, style_names = "myred")

}
