% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_workbook.R
\name{wb_new_workbook}
\alias{wb_new_workbook}
\title{Create a new workbook and add named work sheets}
\usage{
wb_new_workbook(
  sheet_names = NULL,
  ...,
  default = "Sheet",
  creator = NULL,
  title = NULL,
  subject = NULL,
  category = NULL,
  datetime_created = Sys.time(),
  theme = NULL,
  keywords = NULL,
  properties = NULL
)
}
\arguments{
\item{sheet_names}{Optional character vector of worksheet names.}

\item{...}{
  Arguments passed on to \code{\link[openxlsx2:wb_add_worksheet]{openxlsx2::wb_add_worksheet}}
  \describe{
    \item{\code{grid_lines}}{A logical. If \code{FALSE}, the worksheet grid lines will be
hidden.}
    \item{\code{row_col_headers}}{A logical. If \code{FALSE}, the worksheet colname and rowname will be
hidden.}
    \item{\code{tab_color}}{Color of the sheet tab. A  \code{\link[openxlsx2:wb_color]{wb_color()}},  a valid color (belonging to
\code{grDevices::colors()}) or a valid hex color beginning with "#".}
    \item{\code{zoom}}{The sheet zoom level, a numeric between 10 and 400 as a
percentage. (A zoom value smaller than 10 will default to 10.)}
    \item{\code{header,odd_header,even_header,first_header,footer,odd_footer,even_footer,first_footer}}{Character vector of length 3 corresponding to positions left, center,
right.  \code{header} and \code{footer} are used to default additional arguments.
Setting \code{even}, \code{odd}, or \code{first}, overrides \code{header}/\code{footer}. Use \code{NA} to
skip a position.}
    \item{\code{visible}}{If \code{FALSE}, sheet is hidden else visible.}
    \item{\code{has_drawing}}{If \code{TRUE} prepare a drawing output (TODO does this work?)}
    \item{\code{paper_size}}{An integer corresponding to a paper size. See \code{\link[openxlsx2:wb_page_setup]{wb_page_setup()}} for
details.}
    \item{\code{orientation}}{One of "portrait" or "landscape"}
    \item{\code{hdpi,vdpi}}{Horizontal and vertical DPI. Can be set with \code{options("openxlsx2.dpi" = X)},
\code{options("openxlsx2.hdpi" = X)} or \code{options("openxlsx2.vdpi" = X)}}
  }}

\item{default}{Default prefix to use for numbered sheets. Default values are
used if \code{sheet_names = NULL} or if \code{n_sheets} is greater than the length of
\code{sheet_names}. Defaults to "Sheet".}

\item{creator}{Creator of the workbook (your name). Defaults to login username or \code{options("openxlsx2.creator")} if set.}

\item{title, subject, category, keywords}{Additional arguments passed to
\code{\link[openxlsx2:wb_workbook]{openxlsx2::wb_workbook()}}.}

\item{datetime_created}{The time of the workbook is created}

\item{theme}{Optional theme identified by string or number.
See \strong{Details} for options.}

\item{properties}{A named list (typically from
\code{\link[openxlsx2:properties-wb]{openxlsx2::wb_get_properties()}}) used to set new workbook properties for
any values set to \code{NULL}. \code{datetime_created} defaults to \code{\link[=Sys.time]{Sys.time()}} so
must be set to \code{NULL} to inherit value from \code{properties}.}
}
\value{
A \code{wbWorkbook} object.
}
\description{
\code{\link[=wb_new_workbook]{wb_new_workbook()}} is a convenience function wrapping
\code{\link[openxlsx2:wb_workbook]{openxlsx2::wb_workbook()}} and \code{\link[openxlsx2:wb_add_worksheet]{openxlsx2::wb_add_worksheet()}} to offer quick
and easy setup for new workbooks.
}
\examples{
wb_new_workbook()

wb_new_workbook("Sheet 1")

wb_new_workbook(c("Data", "Analysis"))

}
