% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_protect.R
\name{wb_protect_worksheets}
\alias{wb_protect_worksheets}
\title{Protect one or more sheets in a workbook}
\usage{
wb_protect_worksheets(
  wb,
  sheet = NULL,
  protect = TRUE,
  password = NULL,
  properties = c("insertColumns", "insertRows", "deleteColumns", "deleteRows")
)
}
\arguments{
\item{wb}{A workbook object}

\item{sheet}{A name or index of a worksheet}

\item{protect}{Whether to protect or unprotect the sheet (default=TRUE)}

\item{password}{(optional) password required to unprotect the worksheet}

\item{properties}{Defaults to \code{c("insertColumns", "insertRows", "deleteColumns", "deleteRows")}}
}
\value{
A \code{wbWorkbook} object.
}
\description{
Protect one or more sheets in a workbook
}
\examples{
wb <- as_wb(mtcars[1:3, ])

wb_protect_worksheets(wb)

}
