% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_rename.R
\name{wb_rename_data}
\alias{wb_rename_data}
\title{Rename column names in workbook data}
\usage{
wb_rename_data(wb, ..., sheet = 1, start_row = 1, start_col = 1)
}
\value{
A \code{wbWorkbook} object.
}
\description{
Use \code{\link[tidyselect:eval_rename]{tidyselect::eval_rename()}} to rename columns in workbook data.
}
\examples{
wb <- as_wb(mtcars[1:3, c("mpg", "hp")])

wb_rename_data(wb, MPG = mpg, HP = hp)

}
