% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_rename.R
\name{wb_rename_sheets}
\alias{wb_rename_sheets}
\alias{wb_rename_sheets_with}
\title{Rename workbook sheets using a tidyselect syntax}
\usage{
wb_rename_sheets(wb, ...)

wb_rename_sheets_with(wb, .fn, .sheets = tidyselect::everything())
}
\value{
A \code{wbWorkbook} object.

A \code{wbWorkbook} object.
}
\description{
\code{wb_rename_sheets()} and \code{wb_rename_sheets_with()} use the tidyselect package
to rename sheets.
}
\examples{
wb <- as_wb(list(a = mtcars[1:2, ], b = mtcars[3:4, ]))
wb_rename_sheets(wb, sales = "a", inventory = "b")

wb2 <- as_wb(list(a = mtcars[1:2, ], b = mtcars[3:4, ]))
wb_rename_sheets_with(wb2, \(x) paste0("sheet_", x))

}
