% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_rename.R
\name{wb_rename_sheets}
\alias{wb_rename_sheets}
\alias{wb_rename_sheets_with}
\title{Rename workbook sheets using a tidyselect syntax}
\usage{
wb_rename_sheets(wb, ...)

wb_rename_sheets_with(wb, .fn, .sheets = tidyselect::everything())
}
\description{
\code{wb_rename_sheets()} and \code{wb_rename_sheets_with()} use the tidyselect package
to rename sheets.
}
