% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_ext.R
\name{wb_save_ext}
\alias{wb_save_ext}
\title{Save a workboook object to file while filling file name from assigned
workbook title}
\usage{
wb_save_ext(wb, file = NULL, ...)
}
\arguments{
\item{wb}{A \code{wbWorkbook} object to write to file}

\item{file}{A path to save the workbook to}
}
\description{
\code{\link[=wb_save_ext]{wb_save_ext()}} is a helper function that fills in the file name when saving
based on the XSLX title. This function is not stable and may change in the
future.
}
\examples{

library(openxlsx2)

withr::with_tempdir({
  wb_workbook(
    title = "Title used for output file"
  ) |>
    wb_add_worksheet() |>
    wb_save_ext()

  fs::dir_ls()
})

}
