% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_save.R
\name{wb_save_ext}
\alias{wb_save_ext}
\title{Save a workboook object to file while filling file name from assigned
workbook title}
\usage{
wb_save_ext(wb, file = NULL, overwrite = TRUE, ...)
}
\arguments{
\item{wb}{A \code{wbWorkbook} object to write to file}

\item{file}{A path to save the workbook to}

\item{overwrite}{If \code{FALSE}, will not overwrite when \code{file} already exists.}

\item{...}{
  Arguments passed on to \code{\link[openxlsx2:wb_save]{openxlsx2::wb_save}}
  \describe{
    \item{\code{path}}{Deprecated argument. Please use \code{file} in new code.}
  }}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{\link[=wb_save_ext]{wb_save_ext()}} is a helper function extending \code{\link[openxlsx2:wb_save]{openxlsx2::wb_save()}} by
filling a missing file name with the workbook title and validating the file
extension. This function is not stable and may change in the future.
}
\examples{

library(openxlsx2)

withr::with_tempdir({
  wb_workbook(
    title = "Title used for output file"
  ) |>
    wb_add_worksheet() |>
    wb_save_ext()

  fs::dir_ls()
})

}
