% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_save.R
\name{wb_save_ext}
\alias{wb_save_ext}
\title{Save a workboook object to file while filling file name from assigned
workbook title}
\usage{
wb_save_ext(wb, file = NULL, overwrite = TRUE, flush = FALSE)
}
\arguments{
\item{wb}{A \code{wbWorkbook} object to write to file}

\item{file}{A path to save the workbook to}

\item{overwrite}{If \code{FALSE}, will not overwrite when \code{file} already exists.}

\item{flush}{Experimental, streams the worksheet file to disk}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{\link[=wb_save_ext]{wb_save_ext()}} is a helper function extending \code{\link[openxlsx2:wb_save]{openxlsx2::wb_save()}} by
filling a missing file name with the workbook title and validating the file
extension. This function is not stable and may change in the future.
}
\examples{

withr::with_tempdir({
  wb <- wb_new_workbook(
    title = "Title used for output file",
    sheet_name = "Sheet 1"
  )

  wb_save_ext(wb)

  fs::dir_ls()
})

}
