% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dims.R
\name{wb_sheets_cols_to_dims}
\alias{wb_sheets_cols_to_dims}
\alias{wb_cols_to_index}
\alias{wb_cols_to_dims}
\title{Get a list of dimensions from one or more sheets}
\usage{
wb_sheets_cols_to_dims(
  wb,
  cols,
  sheets = NULL,
  start_row = NULL,
  start_col = NULL
)

wb_cols_to_index(
  wb,
  cols,
  sheet = current_sheet(),
  start_row = NULL,
  start_col = NULL
)

wb_cols_to_dims(
  wb,
  cols,
  sheet = current_sheet(),
  start_row = NULL,
  start_col = NULL
)
}
\arguments{
\item{cols}{A character vector or list of character vectors. Length of cols
is recycled to match length of sheets.}

\item{sheets}{Default to use all workbook sheets.}
}
\value{
A list of dims character strings, one per sheet.

An integer vector of column positions matching \code{cols}.

A dims character string covering \code{cols}.
}
\description{
\code{\link[=wb_sheets_cols_to_dims]{wb_sheets_cols_to_dims()}} gets a list of dimension specifications based on
columns. Used to get dimensions for multiple column ranges across multiple
sheets.
}
\examples{
wb <- as_wb(mtcars[1:5, ])

wb_cols_to_index(wb, cols = c("mpg", "hp"), sheet = "Sheet 1")

wb_cols_to_dims(wb, cols = c("mpg", "hp"), sheet = "Sheet 1")

wb_sheets_cols_to_dims(wb, cols = c("mpg", "hp"), sheets = "Sheet 1")

}
\keyword{internal}
