% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_sheets_fmt.R
\name{wb_sheets_fmt}
\alias{wb_sheets_fmt}
\title{Apply formatting function that supports sheet and dims argument to one or
more sheets}
\usage{
wb_sheets_fmt(
  wb,
  .f = openxlsx2::wb_add_numfmt,
  cols = NULL,
  sheets = NULL,
  dims = NULL,
  ...
)
}
\arguments{
\item{.f}{Formatting function to apply to workbook.}

\item{cols}{A character vector or list of character vectors. Length of cols
is recycled to match length of sheets.}

\item{sheets}{Default to use all workbook sheets.}
}
\value{
A \code{wbWorkbook} object.
}
\description{
\code{\link[=wb_sheets_fmt]{wb_sheets_fmt()}} applies a formatting function to one or more sheets and
columns. \code{sheets} is recycled to match the length of \code{dims}. \code{dims} can be
set based on \code{cols} and \code{sheets} or set directly.
}
\keyword{internal}
