% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_split.R
\name{wb_split}
\alias{wb_split}
\title{Use \code{dplyr::group_split} to split a workbook into a list of workbooks}
\usage{
wb_split(file, ..., .key = NULL)
}
\arguments{
\item{file}{An xlsx file, \link[openxlsx2]{wbWorkbook} object or URL to xlsx file.}

\item{...}{
  Arguments passed on to \code{\link[=wb_to_df_list]{wb_to_df_list}}
  \describe{
    \item{\code{sheet_names}}{Character vector of sheet names. If not supplied, all
sheet names from the supplied workbook are used.}
  }}

\item{.key}{Passed to \code{\link[dplyr:group_split]{dplyr::group_split()}}.}
}
\value{
A list of wbWorkbook objects.
}
\description{
\code{\link[=wb_split]{wb_split()}} uses \code{\link[=wb_to_df_list]{wb_to_df_list()}} to extract the data frames from each
sheet of a workbook and then split the data frames by a \code{.key} argument then
convert each new list of data frames back into a wbWorkbook object.
}
\examples{

wb <- as_wb(list(mtcars[1:3,], mtcars[4:6,]))

wb_split(wb, .key = carb)

}
