% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_split.R
\name{wb_to_df_list}
\alias{wb_to_df_list}
\title{Convert a workbook to a list of data frames}
\usage{
wb_to_df_list(wb, sheet_names = NULL, ...)
}
\arguments{
\item{sheet_names}{Character vector of sheet names. If not supplied, all
sheet names from the supplied workbook are used.}

\item{...}{
  Arguments passed on to \code{\link[openxlsx2:wb_to_df]{openxlsx2::wb_to_df}}
  \describe{
    \item{\code{file}}{An xlsx file, \link[openxlsx2]{wbWorkbook} object or URL to xlsx file.}
    \item{\code{sheet}}{Either sheet name or index. When missing the first sheet in the workbook is selected.}
    \item{\code{start_row}}{first row to begin looking for data.}
    \item{\code{start_col}}{first column to begin looking for data.}
    \item{\code{row_names}}{If \code{TRUE}, the first col of data will be used as row names.}
    \item{\code{col_names}}{If \code{TRUE}, the first row of data will be used as column names.}
    \item{\code{skip_empty_rows}}{If \code{TRUE}, empty rows are skipped.}
    \item{\code{skip_empty_cols}}{If \code{TRUE}, empty columns are skipped.}
    \item{\code{skip_hidden_rows}}{If \code{TRUE}, hidden rows are skipped.}
    \item{\code{skip_hidden_cols}}{If \code{TRUE}, hidden columns are skipped.}
    \item{\code{rows}}{A numeric vector specifying which rows in the xlsx file to read.
If \code{NULL}, all rows are read.}
    \item{\code{cols}}{A numeric vector specifying which columns in the xlsx file to read.
If \code{NULL}, all columns are read.}
    \item{\code{detect_dates}}{If \code{TRUE}, attempt to recognize dates and perform conversion.}
    \item{\code{na.strings}}{A character vector of strings which are to be interpreted as \code{NA}.
Blank cells will be returned as \code{NA}.}
    \item{\code{na.numbers}}{A numeric vector of digits which are to be interpreted as \code{NA}.
Blank cells will be returned as \code{NA}.}
    \item{\code{fill_merged_cells}}{If \code{TRUE}, the value in a merged cell is given to all cells within the merge.}
    \item{\code{dims}}{Character string of type "A1:B2" as optional dimensions to be imported.}
    \item{\code{show_formula}}{If \code{TRUE}, the underlying Excel formulas are shown.}
    \item{\code{convert}}{If \code{TRUE}, a conversion to dates and numerics is attempted.}
    \item{\code{types}}{A named numeric indicating, the type of the data.
Names must match the returned data. See \strong{Details} for more.}
    \item{\code{named_region}}{Character string with a \code{named_region} (defined name or table).
If no sheet is selected, the first appearance will be selected. See \code{\link[openxlsx2:wb_get_named_regions]{wb_get_named_regions()}}}
    \item{\code{keep_attributes}}{If \code{TRUE} additional attributes are returned.
(These are used internally to define a cell type.)}
    \item{\code{check_names}}{If \code{TRUE} then the names of the variables in the data frame are checked to ensure that they are syntactically valid variable names.}
    \item{\code{show_hyperlinks}}{If \code{TRUE} instead of the displayed text, hyperlink targets are shown.}
  }}
}
\value{
A list of data frame lists.
}
\description{
\code{\link[=wb_to_df_list]{wb_to_df_list()}} uses \code{\link[openxlsx2:wb_to_df]{openxlsx2::wb_to_df()}} to extract each sheet of a
workbook object into a data frame. Additional parameters \code{...} are recycled
to match the length of sheet names.
}
